chore: moves survey follow ups to community (#4476)
