+++ apps/web/app/api/(internal)/pipeline/route.ts
@@ -3,8 +3,8 @@ import { sendSurveyFollowUps } from "@/app/api/(internal)/pipeline/lib/survey-fo
 import { responses } from "@/app/lib/api/response";
 import { transformErrorToDetails } from "@/app/lib/api/validator";
 import { getIsAIEnabled } from "@/modules/ee/license-check/lib/utils";
-import { getSurveyFollowUpsPermission } from "@/modules/ee/license-check/lib/utils";
 import { sendResponseFinishedEmail } from "@/modules/email";
+import { getSurveyFollowUpsPermission } from "@/modules/survey-follow-ups/lib/utils";
 import { headers } from "next/headers";
 import { prisma } from "@formbricks/database";
 import { cache } from "@formbricks/lib/cache";
