+++ apps/web/modules/survey-follow-ups/components/follow-up-item.tsx
@@ -1,4 +1,4 @@
-import { FollowUpModal } from "@/modules/ee/survey-follow-ups/components/follow-up-modal";
+import { FollowUpModal } from "@/modules/survey-follow-ups/components/follow-up-modal";
 import { Badge } from "@/modules/ui/components/badge";
 import { Button } from "@/modules/ui/components/button";
 import { ConfirmationModal } from "@/modules/ui/components/confirmation-modal";
