+++ apps/web/modules/survey-follow-ups/components/follow-ups-view.tsx
@@ -1,5 +1,5 @@
-import { FollowUpItem } from "@/modules/ee/survey-follow-ups/components/follow-up-item";
-import { FollowUpModal } from "@/modules/ee/survey-follow-ups/components/follow-up-modal";
+import { FollowUpItem } from "@/modules/survey-follow-ups/components/follow-up-item";
+import { FollowUpModal } from "@/modules/survey-follow-ups/components/follow-up-modal";
 import { Button } from "@/modules/ui/components/button";
 import { LockIcon, MailIcon } from "lucide-react";
 import { useTranslations } from "next-intl";
